# tests/GalaxyData.Web.Tests/Hosting/Snapshots/HostTests.TheOpenApiDocumentDescribesTheApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GalaxyData",
    "description": "Browse, query and change data across databases and folders of workbooks. Errors are problem details (RFC 9457) with a code.",
    "version": "v1"
  },
  "paths": {
...
    "/api/browse/page": {
      "post": {
        "tags": [
          "Browse"
        ],
        "summary": "A page of rows to browse, with their columns and count when asked for",
        "operationId": "BrowsePage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowsePageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowsePageDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/browse/trail": {
      "post": {
        "tags": [
          "Browse"
        ],
        "summary": "The crumbs of a trail as they stand: the entities they reach, and the rows chosen in them",
        "operationId": "BrowseTrail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowseTrailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseTrailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrowseFromDto": {
        "required": [
          "entity",
          "key"
        ],
        "type": "object",
        "properties": {
          "entity": {
            "type": "string"
          },
          "key": {
            "type": "array"
          }
        }
      },
      "BrowsePageDto": {
        "required": [
          "queryText",
          "parameters",
          "entity",
          "schema",
          "rows",
          "offset",
          "hasMore",
          "total"
        ],
        "type": "object",
        "properties": {
          "queryText": {
            "type": "string"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryParameterDto"
            }
          },
          "entity": {
            "type": "string"
          },
          "schema": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BrowseSchemaDto"
              }
            ]
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridRowDto"
            }
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "hasMore": {
            "type": "boolean"
          },
          "total": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          }
        }
      },
      "BrowsePageRequest": {
        "required": [
          "source"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/BrowseSourceDto"
          },
          "grid": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GridStateDto"
              }
            ]
          },
          "includeSchema": {
            "type": "boolean",
            "default": true
          },
          "includeCount": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "BrowseSchemaDto": {
        "required": [
          "entity",
          "key",
          "capabilities",
          "columns",
          "references",
          "collections"
        ],
        "type": "object",
        "properties": {
          "entity": {
            "type": "string"
          },
          "key": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesDto"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridColumnDto"
            }
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridReferenceDto"
            }
          },
          "collections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridCollectionDto"
            }
          }
        }
      },
      "BrowseSourceDto": {
        "type": "object",
        "properties": {
          "entity": {
            "type": [
              "null",
              "string"
            ]
          },
          "from": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BrowseFromDto"
              }
            ]
          },
          "navigation": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "BrowseTrailDto": {
        "required": [
          "crumbs"
        ],
        "type": "object",
        "properties": {
          "crumbs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrailStepDto"
            }
          }
        }
      },
      "BrowseTrailRequest": {
        "required": [
          "crumbs"
        ],
        "type": "object",
        "properties": {
          "crumbs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrailCrumbDto"
            }
          }
        }
      },
      "CapabilitiesDto": {
        "required": [
          "canInsert",
          "canUpdate",
          "canDelete"
        ],
        "type": "object",
        "properties": {
          "canInsert": {
            "type": "boolean"
          },
          "canUpdate": {
            "type": "boolean"
          },
          "canDelete": {
            "type": "boolean"
          },
          "insertReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "changeReason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "GridCollectionDto": {
        "required": [
          "navigation",
          "target",
          "multiplicity"
        ],
        "type": "object",
        "properties": {
          "navigation": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "multiplicity": {
            "$ref": "#/components/schemas/Multiplicity"
          }
        }
      },
      "GridColumnDto": {
        "required": [
          "name",
          "type",
          "isKey",
          "canUpdate",
          "insert",
          "readOnlyReason",
          "lineage",
          "reference"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TypeDto"
          },
          "isKey": {
            "type": "boolean"
          },
          "canUpdate": {
            "type": "boolean"
          },
          "insert": {
            "$ref": "#/components/schemas/InsertMode"
          },
          "readOnlyReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "lineage": {
            "$ref": "#/components/schemas/LineageDto"
          },
          "reference": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
      "GridConditionDto": {
        "required": [
          "op"
        ],
        "type": "object",
        "properties": {
          "op": {
            "$ref": "#/components/schemas/GridOp"
          },
          "value": { },
          "valueTo": { }
        }
      },
      "GridFilterDto": {
        "required": [
          "column",
          "conditions"
        ],
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridConditionDto"
            }
          },
          "any": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "GridOp": {
        "enum": [
          "eq",
          "ne",
          "lt",
          "le",
          "gt",
          "ge",
          "between",
          "contains",
          "notContains",
          "startsWith",
          "endsWith",
          "blank",
          "notBlank"
        ]
      },
      "GridReferenceDto": {
        "required": [
          "navigation",
          "target",
          "columns",
          "multiplicity"
        ],
        "type": "object",
        "properties": {
          "navigation": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "multiplicity": {
            "$ref": "#/components/schemas/Multiplicity"
          }
        }
      },
      "GridRowDto": {
        "required": [
          "id",
          "k",
          "v",
          "r"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": [
              "null",
              "string"
            ]
          },
          "k": {
            "type": [
              "null",
              "array"
            ]
          },
          "v": {
            "type": "array"
          },
          "r": {
            "type": [
              "null",
              "array"
            ]
          }
        }
      },
      "GridSortDto": {
        "required": [
          "column"
        ],
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "desc": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "GridStateDto": {
        "type": "object",
        "properties": {
          "filters": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/GridFilterDto"
            }
          },
          "where": {
            "maxLength": 10000,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "sort": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/GridSortDto"
            }
          },
          "offset": {
            "maximum": 9223372036854776000,
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "limit": {
            "maximum": 100000,
            "minimum": 1,
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "InsertMode": {
        "enum": [
          "required",
          "optional",
          "never"
        ]
      },
...
      "LineageDto": {
        "required": [
          "kind",
          "sources",
          "expression"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/LineageKind"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineageSourceDto"
            }
          },
          "expression": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LineageKind": {
        "enum": [
          "direct",
          "computed",
          "aggregated",
          "constant",
          "union",
          "unknown"
        ]
      },
      "LineageSourceDto": {
        "required": [
          "column",
          "path"
        ],
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "path": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "Multiplicity": {
        "enum": [
          "one",
          "zeroOrOne",
          "many"
        ]
      },
...
      "ProblemDetails": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "What the problem is, for clients to tell problems apart by"
          },
          "traceId": {
            "type": "string",
            "description": "The request's trace, to find it in the logs"
          }
        }
      },
...
      "QueryParameterDto": {
        "required": [
          "name",
          "type",
          "value"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": { }
        }
      },
...
      "ScalarKind": {
        "enum": [
          "unknown",
          "boolean",
          "int16",
          "int32",
          "int64",
          "decimal",
          "single",
          "double",
          "string",
          "binary",
          "guid",
          "date",
          "time",
          "dateTime",
          "dateTimeOffset",
          "interval",
          "json"
        ]
      },
...
      "TrailCrumbDto": {
        "type": "object",
        "properties": {
          "entity": {
            "type": [
              "null",
              "string"
            ]
          },
          "navigation": {
            "type": [
              "null",
              "string"
            ]
          },
          "key": {
            "type": [
              "null",
              "array"
            ]
          }
        }
      },
      "TrailStepDto": {
        "required": [
          "entity",
          "label",
          "title",
          "found",
          "problem"
        ],
        "type": "object",
        "properties": {
          "entity": {
            "type": [
              "null",
              "string"
            ]
          },
          "label": {
            "type": "string"
          },
          "title": { },
          "found": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "problem": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TypeDto": {
        "required": [
          "kind",
          "nullable",
          "text"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ScalarKind"
          },
          "nullable": {
            "type": "boolean"
          },
          "text": {
            "type": "string"
          },
          "precision": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "scale": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "length": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "isAnsi": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
    }
  },
...
}